User Registration, User Login
Many file changes for formatting, need to implement Prettier to format
the all the source code.

This looks good so far and is working as expected. There might be some
issues but I'll dig into them later. Very unfortunate that we don't get
specific errors for mysql without using promisify or checking for all of
them. I could write something to handle them all but it seems like to
much work.

closes #4, closes #5